third_party/blink/web_tests/wpt_internal/svg/use-count-svg-filter-on-embedded-content.sub.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<svg width="0" height="0">
  <filter id="svg-blur">
    <feGaussianBlur stdDeviation="1" />
  </filter>
</svg>
<style>
  .css-blur {
    filter: blur(1px);
  }
  .svg-blur {
    filter: url(#svg-blur);
  }
</style>

<div id="empty_outer">
  <div id="empty_inner">
  </div>
</div>

<div id="local_outer">
  <div id="local_inner">
    <iframe id="local_iframe" src="resources/empty.html"></iframe>
  </div>
</div>

<div id="remote_outer">
  <div id="remote_inner">
    <iframe id="remote_iframe" src="https://{{hosts[alt][]}}:{{ports[https][0]}}/wpt_internal/svg/resources/empty.html"></iframe>
  </div>
</div>

<div id="plugin_outer">
  <div id="plugin_inner">
    <embed id="plugin_embed" type="application/x-webkit-test-webplugin"></embed>
  </div>
</div>

<iframe id="playground_iframe"></iframe>

<script type=module>
  import {WebFeature} from '/gen/third_party/blink/public/mojom/use_counter/metrics/web_feature.mojom.m.js';

  async function waitForLoad() {
    if (document.readyState === 'complete') {
        return;
    }
    await new Promise(resolve => window.addEventListener('load', resolve));
  }

  function reset() {
    document.getElementById('empty_outer').className = '';
    document.getElementById('empty_inner').className = '';
    document.getElementById('local_outer').className = '';
    document.getElementById('local_inner').className = '';
    document.getElementById('local_iframe').className = '';
    document.getElementById('remote_outer').className = '';
    document.getElementById('remote_inner').className = '';
    document.getElementById('remote_iframe').className = '';
    document.getElementById('plugin_outer').className = '';
    document.getElementById('plugin_inner').className = '';
    document.getElementById('plugin_embed').className = '';
    document.getElementById('playground_iframe').className = '';
    internals.clearUseCounter(document, WebFeature.kSvgFilterPaintedOnLocalFrame);
    internals.clearUseCounter(document, WebFeature.kSvgFilterPaintedOnRemoteFrame);
    internals.clearUseCounter(document, WebFeature.kSvgFilterPaintedOnWebPlugin);
  }

  async function waitForRender() {
    await new Promise(requestAnimationFrame);
    await new Promise(setTimeout);
  }

  async function getIframeResults(url) {
    document.getElementById('playground_iframe').setAttribute('src', url);
    return await new Promise(resolve => window.addEventListener('message', (e) => {
      resolve(e.data);
    }, {once: true}));
  }

  function expectCounters(local, remote, plugin, context) {
    assert_equals(
      printCounters(
        internals.isUseCounted(document, WebFeature.kSvgFilterPaintedOnLocalFrame),
        internals.isUseCounted(document, WebFeature.kSvgFilterPaintedOnRemoteFrame),
        internals.isUseCounted(document, WebFeature.kSvgFilterPaintedOnWebPlugin)
      ),
      printCounters(local, remote, plugin),
      context
    );
  }

  function expectIframeCounters(message, local, remote, plugin, context) {
    assert_equals(
      printCounters(message.local, message.remote, message.plugin),
      printCounters(local, remote, plugin),
      context
    );
  }

  function printCounters(local, remote, plugin) {
    return `Local: ${local}, Remote: ${remote}, Plugin: ${plugin}`;
  }

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('empty_outer').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Outer filter.');

    reset();
    document.getElementById('empty_inner').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Inner filter.');
  }, 'SVG filter applied to empty divs');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('empty_outer').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Outer filter.');

    reset();
    document.getElementById('empty_inner').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Inner filter.');
  }, 'CSS filter applied to empty divs');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('empty_outer').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Outer filter CSS.');

    document.getElementById('empty_outer').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Outer filter SVG.');

    reset();
    document.getElementById('empty_inner').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Inner filter CSS.');

    document.getElementById('empty_inner').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Inner filter SVG.');
  }, 'CSS & SVG filter applied to empty divs');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('local_outer').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/true, /*remote=*/false, /*plugin=*/false, 'Outer filter.');

    reset();
    document.getElementById('local_inner').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/true, /*remote=*/false, /*plugin=*/false, 'Inner filter.');

    reset();
    document.getElementById('local_iframe').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/true, /*remote=*/false, /*plugin=*/false, 'Iframe filter.');
  }, 'SVG filter applied to local frame');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('local_outer').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Outer filter.');

    reset();
    document.getElementById('local_inner').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Inner filter.');

    reset();
    document.getElementById('local_iframe').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Iframe filter.');
  }, 'CSS filter applied to local frame');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('local_outer').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Outer filter CSS.');

    document.getElementById('local_outer').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/true, /*remote=*/false, /*plugin=*/false, 'Outer filter SVG.');

    reset();
    document.getElementById('local_inner').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Inner filter CSS.');

    document.getElementById('local_inner').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/true, /*remote=*/false, /*plugin=*/false, 'Inner filter SVG.');

    reset();
    document.getElementById('local_iframe').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Iframe filter CSS.');

    document.getElementById('local_iframe').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/true, /*remote=*/false, /*plugin=*/false, 'Iframe filter SVG.');
  }, 'CSS & SVG filter applied to local frame');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('remote_outer').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/true, /*plugin=*/false, 'Outer filter.');

    reset();
    document.getElementById('remote_inner').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/true, /*plugin=*/false, 'Inner filter.');

    reset();
    document.getElementById('remote_iframe').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/true, /*plugin=*/false, 'Iframe filter.');
  }, 'SVG filter applied to remote frame');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('remote_outer').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Outer filter.');

    reset();
    document.getElementById('remote_inner').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Inner filter.');

    reset();
    document.getElementById('remote_iframe').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Iframe filter.');
  }, 'CSS filter applied to remote frame');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('remote_outer').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Outer filter CSS.');

    document.getElementById('remote_outer').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/true, /*plugin=*/false, 'Outer filter SVG.');

    reset();
    document.getElementById('remote_inner').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Inner filter CSS.');

    document.getElementById('remote_inner').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/true, /*plugin=*/false, 'Inner filter SVG.');

    reset();
    document.getElementById('remote_iframe').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Iframe filter CSS.');

    document.getElementById('remote_iframe').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/true, /*plugin=*/false, 'Iframe filter SVG.');
  }, 'CSS & SVG filter applied to remote frame');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('plugin_outer').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/true, 'Outer filter.');

    reset();
    document.getElementById('plugin_inner').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/true, 'Inner filter.');

    reset();
    document.getElementById('plugin_embed').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/true, 'Plugin filter.');
  }, 'SVG filter applied to plugin frame');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('plugin_outer').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Outer filter.');

    reset();
    document.getElementById('plugin_inner').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Inner filter.');

    reset();
    document.getElementById('plugin_embed').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Plugin filter.');
  }, 'CSS filter applied to plugin frame');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Initial load.');

    reset();
    document.getElementById('plugin_outer').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Outer filter CSS.');

    document.getElementById('plugin_outer').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/true, 'Outer filter SVG.');

    reset();
    document.getElementById('plugin_inner').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Inner filter CSS.');

    document.getElementById('plugin_inner').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/true, 'Inner filter SVG.');

    reset();
    document.getElementById('plugin_embed').className = 'css-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Plugin filter CSS.');

    document.getElementById('plugin_embed').className = 'svg-blur';
    await waitForRender();
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/true, 'Plugin filter SVG.');
  }, 'CSS & SVG filter applied to plugin frame');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    let message = await getIframeResults('resources/nested.sub.html');
    expectIframeCounters(message, /*local=*/false, /*remote=*/false, /*plugin=*/false, 'Unfiltered nested load.');
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Unfiltered load.');

    reset();
    await waitForRender();
    document.getElementById('playground_iframe').className = 'svg-blur';
    message = await getIframeResults('resources/nested.sub.html');
    expectIframeCounters(message, /*local=*/false, /*remote=*/false, /*plugin=*/false, 'Filtered nested load.');
    expectCounters(/*local=*/true, /*remote=*/false, /*plugin=*/false, 'Filtered load.');
  }, 'SVG filter applied to nested, but internally unfiltered, local iframe');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    let message = await getIframeResults('https://{{hosts[alt][]}}:{{ports[https][0]}}/wpt_internal/svg/resources/nested.sub.html');
    expectIframeCounters(message, /*local=*/false, /*remote=*/false, /*plugin=*/false, 'Unfiltered nested load.');
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Unfiltered load.');

    reset();
    await waitForRender();
    document.getElementById('playground_iframe').className = 'svg-blur';
    message = await getIframeResults('https://{{hosts[alt][]}}:{{ports[https][0]}}/wpt_internal/svg/resources/nested.sub.html');
    expectIframeCounters(message, /*local=*/false, /*remote=*/false, /*plugin=*/false, 'Filtered nested load.');
    expectCounters(/*local=*/false, /*remote=*/true, /*plugin=*/false, 'Filtered load.');
  }, 'SVG filter applied to nested, but internally unfiltered, remote iframe');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    let message = await getIframeResults('resources/nested.sub.html?filter');
    expectIframeCounters(message, /*local=*/true, /*remote=*/true, /*plugin=*/true, 'Unfiltered nested load.');
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Unfiltered load.');

    reset();
    await waitForRender();
    document.getElementById('playground_iframe').className = 'svg-blur';
    message = await getIframeResults('resources/nested.sub.html?filter');
    expectIframeCounters(message, /*local=*/true, /*remote=*/true, /*plugin=*/true, 'Filtered nested load.');
    expectCounters(/*local=*/true, /*remote=*/false, /*plugin=*/false, 'Filtered load.');
  }, 'SVG filter applied to nested, internally filtered, local iframe');

  promise_test(async (t) => {
    await waitForLoad();

    reset();
    await waitForRender();
    let message = await getIframeResults('https://{{hosts[alt][]}}:{{ports[https][0]}}/wpt_internal/svg/resources/nested.sub.html?filter');
    expectIframeCounters(message, /*local=*/true, /*remote=*/true, /*plugin=*/true, 'Unfiltered nested load.');
    expectCounters(/*local=*/false, /*remote=*/false, /*plugin=*/false, 'Unfiltered load.');

    reset();
    await waitForRender();
    document.getElementById('playground_iframe').className = 'svg-blur';
    message = await getIframeResults('https://{{hosts[alt][]}}:{{ports[https][0]}}/wpt_internal/svg/resources/nested.sub.html?filter');
    expectIframeCounters(message, /*local=*/true, /*remote=*/true, /*plugin=*/true, 'Filtered nested load.');
    expectCounters(/*local=*/false, /*remote=*/true, /*plugin=*/false, 'Filtered load.');
  }, 'SVG filter applied to nested, internally filtered, remote iframe');
</script>
